Add total spike rate to spike rate plots
This adds the number of the total spike rate for the first cell in each
population to the spike rate plots directly, except for some initial amount of
time (one second by default) so as to ignore initial conditions/transients.
This way, if your cell is persistently spiking at the same frequency for the
whole simulation, then you can look at the number printed to get a quick
estimate for its frequency that gets ignores interfering harmonics in the power
spectrum.

	modified:   matlab/functions/plotspk.m