Fix #60: Add option for non-Timesurfer format data
This adds `timesurfer_flag` usage to `matlab/functions/simstudy.m` so that any
flags passed to `simstudy()` e.g. for batches, will be passed to `runsim()`,
and thus large batch simulation runs can take advantage of the "new"
non-Timesurfer format. I've tested this on the cluster and it seems to work.

 Changes to be committed:
	modified:   matlab/functions/simstudy.m